+++ app/src/main/java/com/rawa/notes/repository/NotesRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface NotesRepository {
     fun notes(): Flow<List<Note>>
-    fun getNote(id: Int): Flow<Note>
+    fun getNote(id: Long): Flow<Note>
     suspend fun addNote(note: Note)
     suspend fun removeNote(note: Note)
 }
